services/python/ai-service/pyproject.toml
[project]
name = "medfund-ai-service"
version = "0.1.0"
description = "MedFund AI/ML Service — adjudication AI, fraud detection, OCR, chatbot, analytics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "aiokafka>=0.12.0",
    "redis>=5.2.0",
    "google-generativeai>=0.8.0",
    "scikit-learn>=1.5.0",
    "xgboost>=2.1.0",
    "pytesseract>=0.3.13",
    "pillow>=11.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
